refactor: move ACP bridge plugin into engine built-ins
- add built-in ACP plugin with client/session-store/tools in packages/engine

- register and export builtInAcpPlugin from engine built-in and root entry

- remove remote-server local ACP implementation and plugin injection
