+++ apps/remote-server/src/core/engine-singleton.ts
@@ -1,5 +1,4 @@
 import { Engine } from 'pulse-coder-engine';
-import { acpBridgeIntegration } from './acp/integration.js';
 import { memoryIntegration } from './memory-integration.js';
 import { worktreeIntegration } from './worktree/integration.js';
 import { cronJobTool } from './tools/cron-job.js';
@@ -17,7 +16,6 @@ import { ptcDemoTools } from './tools/ptc-demo.js';
 export const engine = new Engine({
   enginePlugins: {
     plugins: [
-      acpBridgeIntegration.enginePlugin,
       memoryIntegration.enginePlugin,
       worktreeIntegration.enginePlugin,
     ],
