+++ packages/engine/src/built-in/acp-plugin/index.ts
@@ -1,0 +1,2 @@
+export { builtInAcpPlugin } from './acp-plugin';
+export { default } from './acp-plugin';
