+++ packages/engine/src/index.ts
@@ -26,6 +26,7 @@ export { PluginManager } from './plugin/PluginManager.js';
 // 内置插件导出
 export {
   builtInPlugins,
+  builtInAcpPlugin,
   builtInMCPPlugin,
   builtInSkillsPlugin,
   builtInPlanModePlugin,
